+++ code_examples/workshop/attester/generateAccount.ts
@@ -28,7 +28,7 @@ export async function getAccount(
 if (require.main === module) {
   envConfig()
   cryptoWaitReady().then(() => {
-    const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format })
+    const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format, type: 'sr25519' })
 
     generateAccount(keyring)
       .catch((e) => {
