+++ code_examples/workshop/claimer/generateCredential.ts
@@ -67,7 +67,7 @@ export async function generateCredential(
 // don't execute if this is imported by another file
 if (require.main === module) {
   envConfig()
-  const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format })
+  const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format, type: 'sr25519' })
   const signCallbackForKeyring = (keyring: Keyring): Kilt.SignCallback => {
     return async ({ data, alg, publicKey }) => {
       const address =
