+++ code_examples/workshop/claimer/generateLightDid.ts
@@ -44,7 +44,7 @@ export async function generateLightDid(keyring: Keyring): Promise<{
 if (require.main === module) {
   envConfig()
   cryptoWaitReady().then(() => {
-    const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format })
+    const keyring = new Keyring({ ss58Format: Kilt.Utils.ss58Format, type: 'sr25519' })
 
     generateLightDid(keyring)
       .catch((e) => {
